Run front-end app from mock-server
