+++ .gitignore
@@ -21,3 +21,5 @@ dist/onfido.crossDevice.css.map
 dist/onfido.crossDevice.css
 dist/previewer*
 dist/*~*
+test/mock-server/*.pem
+test/mock-server/frontend/
