+++ src/demo/demo.ejs
@@ -12,7 +12,7 @@
       default-src 'self' https://assets.onfido.com;
       script-src 'self' https://www.woopra.com https://assets.onfido.com https://sentry.io;
       style-src 'self' 'unsafe-inline' https://assets.onfido.com;
-      connect-src blob: <%= process.env.NODE_ENV === 'test' ? 'https://localhost:8081' : '' %> *.onfido.com wss://*.onfido.com *.dev.onfido.xyz wss://*.dev.onfido.xyz https://www.woopra.com https://sentry.io;
+      connect-src blob: <%= process.env.NODE_ENV === 'test' ? 'https://localhost:8080' : '' %> *.onfido.com wss://*.onfido.com *.dev.onfido.xyz wss://*.dev.onfido.xyz https://www.woopra.com https://sentry.io;
       img-src 'self' data: blob: https://lipis.github.io/flag-icon-css/;
       media-src blob:;
       object-src 'self' blob:;
