--- test/mock-server/cert.pem
@@ -1,31 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFXTCCA0WgAwIBAgIJANhjm9EVmeNoMA0GCSqGSIb3DQEBCwUAMEUxCzAJBgNV
-BAYTAkFVMRMwEQYDVQQIDApTb21lLVN0YXRlMSEwHwYDVQQKDBhJbnRlcm5ldCBX
-aWRnaXRzIFB0eSBMdGQwHhcNMTcxMTE1MTUyNDAyWhcNMjcxMTEzMTUyNDAyWjBF
-MQswCQYDVQQGEwJBVTETMBEGA1UECAwKU29tZS1TdGF0ZTEhMB8GA1UECgwYSW50
-ZXJuZXQgV2lkZ2l0cyBQdHkgTHRkMIICIjANBgkqhkiG9w0BAQEFAAOCAg8AMIIC
-CgKCAgEAuVMi1u9TlEpyMRq7vfe5x87wyET9MGIVfQNrzRF1CtZEMOHWmb7Jmz3L
-6ZjEI0df0snRluOC4TSctkIqLogWNBkKXPssfvkhPi13obL9aAc90Sz7zAQ2oJ/Q
-9qvVgeFrbxILDOALErMZkvmW/Bi3ZcVVgVMuZ7Uc5mLQz2I3iW9LDD4F0/Tg4ynK
-Cx1nclIgso1OxUDKGKYFeQZVZphlwhf2vFYyHCsQbCPzqMDhQqyshTrL/o6K29Ld
-/7f0fEyIXIATDfQA7o7LNBijET+dtvpz+YoFxxCmsCnCAK0m7w830z3YeP/jGBTC
-S4QspHMkAwxlP/NwpiHowstL20+js3iAAAfBWb3TKIIzYdZduRddN/sO21YHwRHH
-PSR256kzrwUdigQ50HffWDqE4NCCkO3Z6sDDqiU3pruKzaA7HDUCOwz3bbENhZCG
-2tF49W2VeVA1SUojBZ9scAZ+3t/uxtkQliKVUH+XfESrxOGUssF/1mqhHmQPaUNx
-v9OiAC5Y0bqo8dNJjxRao9k548TsIdh0+p3X7RxfHvpV68AH6Mc5DU0dJV11MfLN
-voaDXQRyuR9FvUXTh6VmFnu4Zbe4ulj0htEn/h8/XfOayQZhuvQTw3ejjmidtD6W
-vAbI+yWEQCjBaJbuiPQF61CtenLl5YYbDOP4ClfnoghFN+EGgp8CAwEAAaNQME4w
-HQYDVR0OBBYEFAccjWRDy0JgyDI/ZPYy/X79jfC1MB8GA1UdIwQYMBaAFAccjWRD
-y0JgyDI/ZPYy/X79jfC1MAwGA1UdEwQFMAMBAf8wDQYJKoZIhvcNAQELBQADggIB
-AAWw4mZeNxvD8Ix+4wydPUNbx+1XfN2Oek+BpxBxYbJPBAg600N5SjyW7Gp0AC4E
-Q5BJ0VWEovyD8gaMRgzhZfGoSewBn3X7xBB+KiWEwm5R218s2VTildRFggjBJJI5
-VIkk7wd6VwtjhePdY4uh867T2inyHxbpTzinYgxFbhwHDZfw+lOBZajqdQRMeuB+
-2s12/QJsh8YODpZSdWcjRX1gboiaCfL4RWsNSiFO12QEL5UwtFYWVMQOc2RmDe8A
-TPONEzgpl0Hix6QG2tHYgnTKilN4pmH9Jk3d6eHWim+TeFNvYua2x9y9LLsr3gf4
-JFkXqqYjBkAqgsrnpJpMyIUOEAgi+XYNI11XaAKuze2u4HvYe6yD8cwGULYIYtZj
-yZmbzW6SnrLRqyX6h60c2nUMaXECVkQAOT3k0t0W2ZmJ1DXGLwa5qh9cBkL9/XQn
-QjC3DJclDrAG757xBAcwC82mXM2Ht7+gK9CnXvq7T7yMrgYVk9bKeZl7t8Bz+fAK
-ecLkXtEAKH28ISrBpdTJrIaWxxT6gPS17qy84n0nua+JgHp7p9IgTOfPbaUKoqpH
-xTqGo60SckgAjgWcUzEW2a7eqrQzkY6egXz9DPbjycrrYY5T8fTe6gMtcN/a/fKY
-S/O+QFxB1BNAbFeXTLsJRBaD2JV+GfVNgR0EhidzM6VD
------END CERTIFICATE-----
