+++ test/mock-server/deps.ts
@@ -4,6 +4,7 @@ export { create, getNumericDate } from 'https://deno.land/x/djwt@v1.9/mod.ts'
 
 export type {
   FormDataBody,
+  Middleware,
   RouterContext,
 } from 'https://deno.land/x/oak@v6.3.2/mod.ts'
 
@@ -11,6 +12,7 @@ export {
   Application,
   Router,
   Status,
+  send,
 } from 'https://deno.land/x/oak@v6.3.2/mod.ts'
 
 export { oakCors } from 'https://deno.land/x/cors@v1.2.1/mod.ts'
