--- test/mock-server/key.pem
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJQwIBADANBgkqhkiG9w0BAQEFAASCCS0wggkpAgEAAoICAQC5UyLW71OUSnIx
-Gru997nHzvDIRP0wYhV9A2vNEXUK1kQw4daZvsmbPcvpmMQjR1/SydGW44LhNJy2
-QiouiBY0GQpc+yx++SE+LXehsv1oBz3RLPvMBDagn9D2q9WB4WtvEgsM4AsSsxmS
-+Zb8GLdlxVWBUy5ntRzmYtDPYjeJb0sMPgXT9ODjKcoLHWdyUiCyjU7FQMoYpgV5
-BlVmmGXCF/a8VjIcKxBsI/OowOFCrKyFOsv+jorb0t3/t/R8TIhcgBMN9ADujss0
-GKMRP522+nP5igXHEKawKcIArSbvDzfTPdh4/+MYFMJLhCykcyQDDGU/83CmIejC
-y0vbT6OzeIAAB8FZvdMogjNh1l25F103+w7bVgfBEcc9JHbnqTOvBR2KBDnQd99Y
-OoTg0IKQ7dnqwMOqJTemu4rNoDscNQI7DPdtsQ2FkIba0Xj1bZV5UDVJSiMFn2xw
-Bn7e3+7G2RCWIpVQf5d8RKvE4ZSywX/WaqEeZA9pQ3G/06IALljRuqjx00mPFFqj
-2TnjxOwh2HT6ndftHF8e+lXrwAfoxzkNTR0lXXUx8s2+hoNdBHK5H0W9RdOHpWYW
-e7hlt7i6WPSG0Sf+Hz9d85rJBmG69BPDd6OOaJ20Ppa8Bsj7JYRAKMFolu6I9AXr
-UK16cuXlhhsM4/gKV+eiCEU34QaCnwIDAQABAoICAQCVA1wIJve1TQb0pOJApWgk
-WSkSXyaWOMe+dXnAskDCLfc2tuCvTtEcmlvTwJo1I7bQ7qYwBU+/dWZsxlbUg2gX
-C+thSDVHA1rPfA+tXRaVBv2Vq+3q5ZhnHxoOvaw3euG/tWpdN+ivaIMvwrC8k8Mj
-ecB+9Dtj+CT8uofbuMWfJHXbqiNVgpkxDvD+bu425gKUjzq/MJFPsG7GGWmhzPhF
-UK98UHKKY7YYve+1ySFXUzCFD/nm/QZ6KtuGOovDRmZ+uhyYI5J8PyT1BrpXrCdM
-caW/CFQM9D6Td4AeuxuLLVld6WSeJmKbglWYEvcyY8KZEOZ4vW3ep3VY685a0XZ2
-Ah/9JqqKc8EX3T2Cb2LzfL+KxnrThTd8YI/DKPenhiXe8GFNEdpm8q2kVsy9QuoT
-HYynC1lIdf3yVgUNPY4IEv4kEprErprkXZtMOgMZW0ywmoVZU7c4a9Cs3zckN/vt
-dX6fs7S/yrV8xqOA3GZ7Paizasc0y+7JfjE0Gfl5OU1lKKfPNEdy3sDeqjILk7c+
-qKWsRar+hmEzSjL/DTA+Q0MTh+sWKo6RJZ1r7LSLbnDJMT6Lk8T3rbhckxy++Iw+
-0UIUUbIO9sjXklaSOeRKX0iXA0jLeHBLpd74W1zShUdNiVapuM+5nSJ2YTsJgR1W
-7ndEcY8uhvJBBpYb+sIMwQKCAQEA3criUnnPAt7tmMtFMyGiuXfw5Ty6hk8zuBSu
-d3FJedyvNhPHEWSrX+aI8dqCiuVSefWqkzhsLao5swK4rjVEPxu9leFb25i/DWBJ
-xW3XMXogD5clLjwjfr48u49XlsR6w/IfNw816zXKUVNYOoakK4tOrboTF/hSJofv
-OIHxXeEK0UnRyAj5NyOyVmVRaT/fwK0oSCqfY/pBDNF9k9nBgn/9I7d1gV8DwWrF
-ABPCpvUUnBe5NKmmOZ1EeHdVRN2DQRZAcPMphhvTJZQizCdrV0gKBxIi8GJWhwLV
-QcejCTnz4PwqLbel/2gfha/fvIrgVyeaZul7NEZBoXu29uMUUwKCAQEA1ehhujd8
-R+/cFzOlfo80VSzeRwjq1BkVah315XB+IefIjVE07DB48DrsQVCcTR5IjyzxmBHZ
-lW1wWzzw58kcKaM4ZkN3fxn5RE76CG6sIviVurNriWFdKF7MB8wbNUQTpPmsbM3e
-k1hyqMNhu2xOEfPTT/PpCtu3yw0/gnVFo3g2RIX9gzHIPbxWsvrdLtk2eRN3fgWV
-TUj0QiIszTzntUSNreSMvNsCwtMDbNxiRPJkpfb13srrAerksItS9XbK6CNtLts+
-p7bExmrlusxrp7290AUgERFBjJ6s7Y33SfxmZOX93BczQ8NwScJwfEJDbPzCs18p
-3hkeNhgdJTjPBQKCAQAhgrv7Vigu+NvRUDUeaqCZgUPGfwx3qpPX5lgN9/HwFaXr
-d8WGHT2/eaeLU0aSEeAKSVICeG3v4F5ja6/XsnijSF+2u5wTa9rB/KBo0hXHhfo+
-IMofWDrMtZWbiSbkwr8ByB1rRkMRTMlKzFryTNGVOrEikpC7+GsBI1zDnu6HDq+B
-u7OUk7aD3ELfxuMsEdiO1hrV/IqkonUdqLFem3Ifc5MdyGRara51lNo7DsFEVJdO
-lk8HY/KXlel4gh9CDm38cY4/Ap5tgtq+PMmdKv+O6nWJZ/Ybd+A2+fiavb0mcbkl
-bzQ4QB02n9bTbiVR6XHH+VXc4r4Ywdy8t2KHnvu9AoIBAAj4IYDparCpghUG+V2B
-K7SWmQtpeE9Ebf1xHCpuuRs62NH0Ib/1mFyaW+WxN5yUNH6sZ8T0mqcFW7y3pXJS
-oHeufGLtWUlfEXCnzxwjZ9xfiAe4Fj5AcuOKdgutp0ZogFpejbhSOY78GnKxHpv5
-QpzGGIo1tAPciqxc9djGihD8rWnaJHMn1zmfeBMxKa0ZNouFIXGLUmNXN8a+IBrK
-/vw+z6IG/0+77zW0G2SpKQWzFD1RSaX55DjEXKp6bRQV0xRAyAc68jmeFEotmUxn
-3eu+rQLX9CxDCPiYfq/HUCMuDi0U57wezOt+emrdWFIdhS5AuvlCeELz0qx5/l7d
-LvECggEBAMeflAhtMhLZ94L/M/H+kY3f9hnCxxB1JjpeOaWfqv0KS5e4BdFZ7a/R
-o6xWNEOP1oi6JyAGfJJebgsx7TkcsdpUq2JGgaOw+nCmeQb4B+Xyhzy/FIKfM5g+
-WGHgYVRCXi1BTKgFdLx1WM+hVZXS/YUXy6t8sCps8EEHAJ9tXwNY8oXs9BrZx9Eo
-kvR3583RXojuUtrbqahaDk7ImsFduaXAxsBMUH+DJrqEfCcpWTkDFFr/pn113Hdl
-68zNGLLIuNj988P9rJXvN/mcUBoZu5Mp6qIBRHP+8rZfM5JX96pku+pp3quzCWz0
-WkBFPuRkNxh6jX7aEmCNLCavnQzgOpk=
------END PRIVATE KEY-----
